backend-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic[email]
passlib
bcrypt>=4.0.0
python-jose[cryptography]
python-multipart
